refactor: extract SFTP directory tree helpers into transfer-sftp-dir module
